[#159880517] refactor dashboard component and add resposiveness
