+++ src/_css/AllocationsComponent.scss
@@ -12,3 +12,9 @@
   position: absolute;
   width: 90px;
 }
+
+@media (max-width: 767px) {
+  .allocations-list {
+    padding: 0 30px;
+  }
+}
