+++ src/_css/UsersComponent.scss
@@ -23,3 +23,9 @@
   display: flex;
   height: 50px;
 }
+
+@media (max-width: 767px) {
+  .users-list {
+    padding: 0 30px;
+  }
+}
