+++ src/components/Dashboard/AnalyticsCardComponent.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Segment, Image } from 'semantic-ui-react';
 import PropTypes from 'prop-types';
 
-import '../_css/DashboardComponent.css';
+import '../../_css/DashboardComponent.css';
 
 const AnalyticsCardComponent = props => (
   <Segment.Group horizontal>
